+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 1.0.6-dev0
+## 1.0.6
 
 ### Fixes
 
 * **Google Drive connector now strips the leading dot in extensions properly**
+* **Google Drive permissions conform to FileData schema**
+* **Confluence permissions conform to FileData schema**
 
 ## 1.0.5